tosec: change structure to prep for new parser
