+++ src/shiratsu-lib/src/parse/tosec/mod.rs
@@ -1,5 +1,6 @@
 mod dat;
 mod filename;
+mod legacy_parser;
 
 pub use dat::*;
 pub use filename::TosecNameable;
